Silence compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@21363 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseMerge/BaseView.cpp
@@ -4074,7 +4074,7 @@ int CBaseView::CountMultiLines( int nLine )
     {
         if (m_ScreenedViewLine[nLine].bSet)
         {
-            return m_ScreenedViewLine[nLine].SubLines.size();
+            return (int)m_ScreenedViewLine[nLine].SubLines.size();
         }
     }
